// src/services/enums/standing-stones.ts
export enum EStandingStones {
    Nothing = 'Не выбрано',
    Warrior = 'Воин',
    Lord = 'Лорд',
    Lady = 'Леди',
    Steed = 'Конь',
    Mage = 'Маг',
    Apprentice = 'Ученик',
    Atronach = 'Атронах',
    Ritual = 'Ритуал',
    Thief = 'Вор',
    Tower = 'Башня',
    Lover = 'Любовник',
    Shadow = 'Тень',
    Serpent = 'Змей'
}