// src/services/enums/aedraes.ts
export enum EAedraes {
    Nothing = 'Не выбрано',
    Julianos = 'Джулианос',
    Zenithar = 'Зенитар',
    Kynareth = 'Кинарет',
    Akatosh = 'Акатош',
    Dibella = 'Дибелла',
    Arkay = 'Аркей',
    Mara = 'Мара',
    Stendarr = 'Стендарр',
    Talos = 'Талос',
    Auriel = 'Ауриель'
}