// src/services/enums/daedraes.ts
export enum EDaedraes {
    Nothing = 'Не выбрано',
    Boethia = 'Боэтия',
    Peryite = 'Периайт',
    Hermeus = 'Хермеус Мора',
    Azura = 'Азура',
    Nocturnal = 'Ноктюрнал',
    Malacath = 'Малакат',
    Mehrunes = 'Мерунес Дагон',
    Mephala = 'Мефала',
    Meridia = 'Меридия',
    Sanguine = 'Сангвин',
    Namira = 'Намира',
    Vaermina = 'Вермина',
    Clavicus = 'Клавикус Вайл',
    Sheogorath = 'Шеогорат'
}